Pubg_Final_Placement_Prediction
In this project the goal is to create the model which predict players finishing placement based on their final stats. 
Train and test csv's are individually used to train and test the data. 
Linear Regression algorithm is used  which gave 94% accuracy.Which on top that Artificial Neural Networks
(MLPRegressor is used to enhance the predictive power of model but the result remains same and had no improvement.